Outer JobScheduleListener / rename JobSuspender
